requirements.txt
gradio==4.44.0
google-generativeai==0.8.2
PyPDF2==3.0.1
python-docx==1.1.2
pathlib
python-dotenv